+++ CHANGELOG.markdown
@@ -6,7 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
-* Fixed wrong expression optimization of OP SUB instruction ([#198](https://github.com/cardillan/mindcode/issues/198)).
+* Fixed wrong definition of `control color` instruction ([#194](https://github.com/cardillan/mindcode/issues/194)).
+* Fixed wrong expression optimization of `op sub` instruction ([#198](https://github.com/cardillan/mindcode/issues/198)).
 
 ## 2.7.3 - 2024-12-28